небольшие изменеия в структуре
crud.py - методы запросов к БД
cud_students.py - методы обрабока запросов ввода информации добавление\удаление\изменение
get_info.py - методы вывода запрошенной информации
