+++ main.py
@@ -1,4 +1,4 @@
-import controller.ctrl as ctrl
+import controller.controller as ctrl
 
 if __name__ == '__main__':
 	ctrl.run()
