--- user_interface.py
@@ -1,53 +1,0 @@
-def main_menu():
-	print('=' * 50)
-	print('Варианты работы программы')
-	print('-' * 50)
-	print('1. Менеджер студентов(Создать/редактировать/удалить)')
-	print('2. Информация о студентах(Все/класс/один)')
-	print('3. Импорт информации')
-	print('4. Выход из программы')
-
-
-def export_menu():
-    print('=' * 50)
-    print('Варианты работы программы')
-    print('-' * 50)
-    print('1. Вывод всех данных')
-    print('2. Вывод данных по указанным фамилии и имени')
-    print('3. Вывод данных по классу')
-    print('4. Вернуться в главное меню')
-    print('5. Выход из программы')
-
-
-def import_menu():
-	print('=' * 50)
-	print('Варианты работы программы')
-	print('-' * 50)
-	print('1. Импортировать в .csv (сохранить как...)')
-	print('2. Вернуться в главное меню')
-	print('3. Выход из программы')
-
-
-def error():
-	print('-' * 50)
-	print('------ Ошибка -------')
-	print('-' * 50)
-
-
-def menu_data_actions():
-    print('=' * 50)
-    print('Варианты работы программы')
-    print('-' * 50)
-    print('1. Создать данные')
-    print('2. Изменить данные')
-    print('3. Удалить данные')
-    print('4. Вернуться в главное меню')
-    print('5. Выход из программы')
-
-
-def exit_menu():
-	print('=' * 50)
-	print('Выберите дальнейшее действие')
-	print('-' * 50)
-	print('1. Выход в главное меню')
-	print('2. Выход из программы')
