+++ user_interface/interface.py
@@ -1,17 +1,31 @@
 def main_menu():
+	print('=' * 50)
+	print('Варианты работы программы')
 	print('-' * 50)
+	print('1. Менеджер студентов(Создать/редактировать/удалить)')
+	print('2. Информация о студентах(Все/класс/один)')
+	print('3. Импорт информации')
+	print('4. Выход из программы')
 
 
-def menu_stdn_manager():
-	print('-' * 50)
-
-
-def menu_stdn_info():
-	print('-' * 50)
+def export_menu():
+    print('=' * 50)
+    print('Варианты работы программы')
+    print('-' * 50)
+    print('1. Вывод всех данных')
+    print('2. Вывод данных по указанным фамилии и имени')
+    print('3. Вывод данных по классу')
+    print('4. Вернуться в главное меню')
+    print('5. Выход из программы')
 
 
 def import_menu():
+	print('=' * 50)
+	print('Варианты работы программы')
 	print('-' * 50)
+	print('1. Импортировать в .csv (сохранить как...)')
+	print('2. Вернуться в главное меню')
+	print('3. Выход из программы')
 
 
 def error():
@@ -20,10 +34,20 @@ def error():
 	print('-' * 50)
 
 
+def menu_data_actions():
+    print('=' * 50)
+    print('Варианты работы программы')
+    print('-' * 50)
+    print('1. Создать данные')
+    print('2. Изменить данные')
+    print('3. Удалить данные')
+    print('4. Вернуться в главное меню')
+    print('5. Выход из программы')
+
+
 def exit_menu():
 	print('=' * 50)
 	print('Выберите дальнейшее действие')
 	print('-' * 50)
 	print('1. Выход в главное меню')
 	print('2. Выход из программы')
-
